Fzz/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
astroid==2.15.4
attrs==22.2.0
bcrypt==4.1.2
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
cleo==2.0.1
cli-helpers==2.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.6
crashtest==0.4.1
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
dill==0.3.6
distlib==0.3.6
dulwich==0.21.3
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.9.0
fonttools==4.33.3
fpdf==1.7.2
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httptools==0.6.1
identify==2.5.23
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.3
jose==1.0.0
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.2
lazy-object-proxy==1.9.0
lockfile==0.12.2
MarkupSafe==2.1.3
matplotlib==3.6.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
Naked==0.1.32
nodeenv==1.7.0
numpy==1.22.3
packaging==23.0
pandas==1.4.2
passlib==1.7.4
pendulum==3.0.0
pexpect==4.8.0
pgcli==4.0.1
pgspecial==2.1.1
Pillow==9.1.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
pre-commit==3.2.2
premailer==3.10.0
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg-pool==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pychattr==0.2.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pylint==2.17.3
pyparsing==3.0.9
pyproject_api==1.5.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.1
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
scipy==1.8.0
setproctitle==1.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.27.0
status==0.2.5
tabulate==0.9.0
time-machine==2.13.0
tomli==2.0.1
tomlkit==0.11.7
tox==4.4.5
tqdm==4.66.1
trove-classifiers==2023.3.9
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.15
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.19.0
watchfiles==0.21.0
